Move file check to servlet-map
